requirements.txt
bidict==0.22.1
click==8.1.3
dnspython==2.3.0
eventlet==0.33.3
Flask==2.2.3
Flask-SocketIO==5.3.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pymongo==4.3.3
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
six==1.16.0
Werkzeug==2.2.3
zope.event==4.6
zope.interface==6.0
